Add: expire cache
