+++ ssm_utils/Cargo.toml
@@ -10,7 +10,9 @@ rust-version = "1.76.0"
 [dependencies]
 aws-config = { version = "1.5.0", features = ["behavior-version-latest"] }
 aws-sdk-ssm = { version = "1.30.0" }
+mini-moka = { version = "0.10.3", optional = true }
 thiserror = { version = "1" }
 
 [features]
-default = []
+ default = ["expire"]
+expire = ["dep:mini-moka"]
