+++ ssm_utils/src/lib.rs
@@ -1,8 +1,7 @@
-use std::{
-    collections::HashMap,
-    sync::{Arc, RwLock, RwLockWriteGuard},
-};
+pub use crate::cache::CachedClient;
+use cache::Cache;
 
+pub mod cache;
 pub mod sdk {
     pub use aws_sdk_ssm::*;
 }
@@ -41,36 +40,12 @@ impl Client {
         Self::from_ssm_client(sdk::Client::from_conf(conf.into()))
     }
 
-    /// SSMの値をキャッシュできるようにします
-    /// ```no_run
-    /// # use ssm_utils::*;
-    ///
-    /// let client = Client::from_env().await.with_cache();
-    /// client.get("aaa").await;
-    /// client.get("aaa").await; // キャッシュされている
-    /// ```
-    pub fn with_cache(self) -> Client<RwCache> {
-        Client {
-            ssm: self.ssm,
-            cache: RwCache::new_cache(),
-        }
-    }
-
     /// Mock用のClientを作ります。
     /// このモードでは、環境変数の値から確認するようになります。
     pub fn mock() -> Client {
         Client {
             ssm: None,
             cache: (),
-        }
-    }
-
-    /// Mock用のClientを作ります。
-    /// このモードでは、mapの中身を確認し、その後環境変数を確認します。
-    pub fn mock_from_map(map: HashMap<String, String>) -> CachedClient {
-        Client {
-            ssm: None,
-            cache: Arc::new(RwLock::new(map)),
         }
     }
 }
@@ -127,64 +102,6 @@ impl<C: Cache> Client<C> {
     /// mockかどうか。
     pub fn is_mock(&self) -> bool {
         self.ssm.is_none()
-    }
-}
-
-/// キャッシュを規定する
-pub trait Cache: Clone {
-    fn new_cache() -> Self
-    where
-        Self: Sized;
-    fn get(&self, key: &str) -> Option<String>;
-    fn set(&self, key: &str, value: &str);
-}
-
-/// キャッシュしない
-impl Cache for () {
-    fn new_cache() -> Self {}
-    /// 必ずNone
-    #[inline]
-    fn get(&self, _key: &str) -> Option<String> {
-        None
-    }
-    /// noop
-    #[inline]
-    fn set(&self, _key: &str, _value: &str) {}
-}
-
-pub type RwCache = Arc<RwLock<HashMap<String, String>>>;
-/// キャッシュ付きssm Client
-pub type CachedClient = Client<RwCache>;
-
-impl Cache for RwCache {
-    fn new_cache() -> Self {
-        Arc::new(RwLock::new(HashMap::new()))
-    }
-    /// キャッシュから取得
-    /// Readのロックがかかるので、ほかにwriteのロックを書けてると待機します。
-    fn get(&self, key: &str) -> Option<String> {
-        self.as_ref()
-            .read()
-            .ok()
-            .and_then(|rg| rg.get(key).cloned())
-    }
-    fn set(&self, key: &str, value: &str) {
-        if let Ok(mut map) = self.write() {
-            map.insert(key.to_owned(), value.to_owned());
-        }
-    }
-}
-
-impl CachedClient {
-    /// キャッシュを取得する
-    /// これがdropされないと[`Client::get`]が待機します。
-    pub fn get_mut_cache(&self) -> Option<RwLockWriteGuard<HashMap<String, String>>> {
-        self.cache.write().ok()
-    }
-
-    /// キャッシュを消す
-    pub fn clear_cache(&self) {
-        self.cache.write().expect("poisoned lock").clear();
     }
 }
